+++ lib/enums/menu_action.dart
@@ -1,0 +1,1 @@
+enum MenuAction { logout }
