+++ lib/services/auth/auth_provider.dart
@@ -3,6 +3,8 @@ import "package:mynotes/services/auth/auth_user.dart";
 abstract class AuthProvider {
   AuthUser? get currentUser;
 
+  Future<void> initialize();
+
   Future<AuthUser> login({
     required String email,
     required String password,
