+++ lib/services/auth/auth_service.dart
@@ -1,9 +1,16 @@
 import "package:mynotes/services/auth/auth_provider.dart";
 import "package:mynotes/services/auth/auth_user.dart";
+import "package:mynotes/services/auth/firebase_auth_provider.dart";
 
 class AuthService implements AuthProvider {
   final AuthProvider provider;
+
   const AuthService(this.provider);
+
+  factory AuthService.firebase() => AuthService(FirebaseAuthProvider());
+
+  @override
+  Future<void> initialize() => provider.initialize();
 
   @override
   Future<AuthUser> createUser({
