+++ lib/services/auth/firebase_auth_provider.dart
@@ -1,3 +1,4 @@
+import "package:firebase_core/firebase_core.dart";
 import "package:mynotes/services/auth/auth_exceptions.dart";
 import "package:mynotes/services/auth/auth_provider.dart";
 import "package:mynotes/services/auth/auth_user.dart";
@@ -5,7 +6,16 @@ import "package:mynotes/services/auth/auth_user.dart";
 import "package:firebase_auth/firebase_auth.dart"
     show FirebaseAuth, FirebaseAuthException;
 
+import "../../firebase_options.dart";
+
 class FirebaseAuthProvider implements AuthProvider {
+  @override
+  Future<void> initialize() async {
+    await Firebase.initializeApp(
+      options: DefaultFirebaseOptions.currentPlatform,
+    );
+  }
+
   @override
   Future<AuthUser> createUser({
     required String email,
