+++ lib/views/login_view.dart
@@ -1,6 +1,7 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:mynotes/constants/routes.dart' as routes;
+import 'package:mynotes/services/auth/auth_exceptions.dart';
+import 'package:mynotes/services/auth/auth_service.dart';
 
 import '../utilities/show_error_dialog.dart';
 // import 'dart:developer' as devtools show log;
@@ -59,13 +60,14 @@ class _LoginViewState extends State<LoginView> {
               final email = _email.text;
               final password = _password.text;
               try {
-                await FirebaseAuth.instance.signInWithEmailAndPassword(
+                await AuthService.firebase().login(
                   email: email,
                   password: password,
                 );
                 if (context.mounted) {
                   final isEmailVerified =
-                      FirebaseAuth.instance.currentUser?.emailVerified ?? false;
+                      AuthService.firebase().currentUser?.isEmailVerified ??
+                          false;
                   final routeTo = isEmailVerified
                       ? routes.notesRoute
                       : routes.verifyEmailRoute;
@@ -74,27 +76,20 @@ class _LoginViewState extends State<LoginView> {
                     (_) => false,
                   );
                 }
-              } on FirebaseAuthException catch (e) {
-                if (e.code == "user-not-found") {
-                  await showErrorDialog(
-                    context,
-                    "User not found",
-                  );
-                } else if (e.code == "wrong-password") {
-                  await showErrorDialog(
-                    context,
-                    "Wrong password",
-                  );
-                } else {
-                  await showErrorDialog(
-                    context,
-                    "Error ${e.code}",
-                  );
-                }
-              } catch (e) {
+              } on UserNotFoundAuthException {
                 await showErrorDialog(
                   context,
-                  e.toString(),
+                  "User not found",
+                );
+              } on WrongPasswordAuthException {
+                await showErrorDialog(
+                  context,
+                  "Wrong password",
+                );
+              } on GenericAuthException {
+                await showErrorDialog(
+                  context,
+                  "Authentication error",
                 );
               }
             },
