+++ lib/views/verify_email_view.dart
@@ -1,6 +1,6 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:mynotes/constants/routes.dart' as routes;
+import 'package:mynotes/services/auth/auth_service.dart';
 
 class VerifyEmailView extends StatefulWidget {
   const VerifyEmailView({super.key});
@@ -23,14 +23,13 @@ class _VerifyEmailViewState extends State<VerifyEmailView> {
             "If you haven't received a verification email yet, press the button below."),
         TextButton(
           onPressed: () async {
-            final user = FirebaseAuth.instance.currentUser;
-            await user?.sendEmailVerification();
+            await AuthService.firebase().sendEmailVerification();
           },
           child: const Text("Resend email verification"),
         ),
         TextButton(
           onPressed: () async {
-            await FirebaseAuth.instance.signOut();
+            await AuthService.firebase().logout();
             if (context.mounted) {
               Navigator.of(context).pushNamedAndRemoveUntil(
                 routes.registerRoute,
